Refactored usage of connect-flash.
Set up middleware that provides the error, warn, info, and log
variables to the view engine, rather than needing to do so in every
place.

Note that we decided there was no point to adding the message to the
session state for GET requests that do not cause a redirect.