Fixed logic to keep track of attempts made.
When the graph is first downloaded, a document containing the team id
and the submission name are stored in a collection. No subsequent
downloads will cause the key to be set in Redis.

During the time period when submissions are accepted, each successful
attempt pushes a timestamp onto an array of the previously mentioned
document.