Added support for additional playback controls.
Includes step-forward, step-backward, and play. Note that a seek can be
performed by changing the step number in the text field.

As of right now, the entire simulation is reapplied, up to that point,
when a step backwards is made; however a "reverse diff" could be
constructed in an online fashion for efficiency.