Modified the layout and responsiveness of the graph view.
Needed to increase the size of the legend in order to get the scores
for each team to display at half-width.

Made it so that the edges are only added if there are few enough nodes.
This is necessary for performance reasons with zooming.

Added support for resizing the window of the browser to cause the graph
to be shifted around, which helps to keep it centered on screen. Would
prefer to not have scaling applied, and only translate.

The uncolored team name is now a variable, defined as $uncolored.